archive/prototypes/avalon_online_py/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
eventlet==0.33.3
discord.py==2.3.2
python-dotenv==1.0.0
email-validator==2.1.0
itsdangerous==2.1.2
Werkzeug==2.3.7
SQLAlchemy==2.0.23
Jinja2==3.1.2
WTForms==3.1.1
psycopg2-binary==2.9.6
